# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Contoso API",
    "description": "Contoso API Information",
    "contact": {
      "name": "Gisela Torres"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/swagger.json": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/invoices": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "description": "Endpoint to return invoices",
        "operationId": "getInvoices",
        "responses": {
          "200": {
            "description": "Invoices returned successfully",
            "schema": {
              "$ref": "#/definitions/Invoices"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Invoices"
        ],
        "description": "",
        "operationId": "createInvoice",
        "parameters": [
          {
            "name": "newInvoice",
            "in": "body",
            "description": "Invoice to be created",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Invoice"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Invoice created successfully",
            "schema": {
              "$ref": "#/definitions/Invoice"
            }
          }
        }
      }
    },
    "/invoices/{id}": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "description": "",
        "operationId": "getInvoiceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Invoice ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice returned successfully",
            "schema": {
              "$ref": "#/definitions/Invoice"
            }
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      },
      "put": {
        "tags": [
          "Invoices"
        ],
        "description": "",
        "operationId": "updateInvoice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Invoice ID"
          },
          {
            "name": "updatedInvoice",
            "in": "body",
            "description": "Invoice to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Invoice"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice updated successfully",
            "schema": {
              "$ref": "#/definitions/Invoice"
            }
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Invoices"
        ],
        "description": "",
        "operationId": "deleteInvoice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Invoice ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice deleted successfully"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      }
    },
    "/accounts": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "description": "Endpoint to return accounts",
        "operationId": "getAccounts",
        "responses": {
          "200": {
            "description": "Accounts returned successfully",
            "schema": {
              "$ref": "#/definitions/Accounts"
            }
          }
        }
      }
    },
    "/support": {
      "get": {
        "tags": [
          "Support"
        ],
        "description": "Endpoint to return support cases",
        "operationId": "getCases",
        "responses": {
          "200": {
            "description": "Cases returned successfully",
            "schema": {
              "$ref": "#/definitions/SupportCases"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Invoice": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "date": {
          "type": "string",
          "example": "2021-05-05"
        },
        "accountName": {
          "type": "string",
          "example": "Franecki Group"
        },
        "contact": {
          "type": "string",
          "example": "Cassidy.Lemke81@yahoo.com"
        },
        "price": {
          "type": "number",
          "example": 380
        },
        "status": {
          "type": "string",
          "example": "Paid"
        }
      }
    },
    "Invoices": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "date": {
            "type": "string",
            "example": "2021-05-05"
          },
          "accountName": {
            "type": "string",
            "example": "Franecki Group"
          },
          "contact": {
            "type": "string",
            "example": "Cassidy.Lemke81@yahoo.com"
          },
          "price": {
            "type": "number",
            "example": 380
          },
          "status": {
            "type": "string",
            "example": "Paid"
          }
        }
      }
    },
    "Accounts": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "accountName": {
            "type": "string",
            "example": "WingTip Toys"
          },
          "primaryContact": {
            "type": "string",
            "example": "Bart Friday"
          },
          "contactEmail": {
            "type": "string",
            "example": "b.friday@wingtiptoys.com"
          }
        }
      }
    },
    "SupportCases": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Lynne Robbins"
          },
          "address": {
            "type": "string",
            "example": "3800 148th Ave NE"
          },
          "phone": {
            "type": "string",
            "example": "425-555-1234"
          },
          "email": {
            "type": "string",
            "example": "lrobbins@proseware.com"
          },
          "notes": {
            "type": "string",
            "example": "Financial Advisory"
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "swagger.json",
    "url": "http://localhost:3000/swagger.json"
  }
}